+++ CHANGELOG.md
@@ -38,6 +38,8 @@ high state of flux, you're at risk of it changing without notice.
     - make `fromNullableK` lazy (@gcanti)
     - make `chainNullableK` lazy (@gcanti)
     - remove `getValidation`, use `getApplicativeValidation` and `getAltValidation` instead (@gcanti)
+    - make `parseJSON` pipeable (@gcanti)
+    - make `stringifyJSON` pipeable (@gcanti)
 
 # 2.9.1